refactor(application-rules): move application rules to a sub dir
