+++ src/applications/index.ts
@@ -2,7 +2,7 @@ export { ApplicationsModule } from './applications.module';
 export { ApplicationsService } from './applications.service';
 export { ApplicationsRepository } from './applications.repository';
 export { ApplicationsController } from './applications.controller';
-export { ApplicationRulesService } from './application-rules.service';
+export { ApplicationRulesService } from './rules/application-rules.service';
 export { CallsRepository } from './calls.repository';
 export { CreateApplicationApi, GetApplicationApi } from './api-docs';
 export { ApplicationDetailDto } from './dto/application-detail.dto';
